Updated health.
Added colors dark green and red. Now the health levels are dark green
(health > 80), green (health > 60), yellow (health > 40), orange
(health > 20), and red (health > 0).

+++ Source/FreedomGeometrics/FreedomGeometricsCharacter.cpp
@@ -225,7 +225,7 @@ int AFreedomGeometricsCharacter::getMaxHealth()
 	return MaxHealth;
 }
 
-void AFreedomGeometricsCharacter::addHealth(float val)
+void AFreedomGeometricsCharacter::addHealth(int val)
 {
 	Health += val;
